Adds .keep file in output_images

+++ utils/clearOutputFiles.js
@@ -5,6 +5,7 @@ const clearOutputFiles = (outputDir) => {
   const outputFiles = fs.readdirSync(outputDir);
   for (const f of outputFiles) {
     try {
+      if (f === ".keep") continue;
       fs.unlinkSync(path.join(outputDir, f));
     } catch (err) {
       console.log(err);
